Move transform package to pkg directory
