+++ pkg/internal/appolly/appolly.go
@@ -17,8 +17,8 @@ import (
 	"github.com/grafana/beyla/pkg/internal/pipe"
 	"github.com/grafana/beyla/pkg/internal/pipe/global"
 	"github.com/grafana/beyla/pkg/internal/request"
-	"github.com/grafana/beyla/pkg/internal/transform"
-	"github.com/grafana/beyla/pkg/internal/transform/kube"
+	"github.com/grafana/beyla/pkg/transform"
+	"github.com/grafana/beyla/pkg/transform/kube"
 )
 
 func log() *slog.Logger {
